requirements.txt
pynput
pyperclip
pyautogui
customtkinter
python-dotenv
groq

